Complete
Add a Search Feature

Prompt the user to input a book title to search for.

If the book is found, display a message indicating it is available.

If the book is not found, display a message that it’s not in the collection.

Limit Borrowing

Add a feature that tracks how many books a user has borrowed.

Limit the number of books to 3 at a time.

Check-in a borrowed book

Add a feature that enables the user to check in a book that’s been checked out.

Check If the book is checked out. If it is remove the checked-out flag to check the book in. If it isn’t, inform the user.

Added additional "test" hidden feature to allow quick testing